pyproject.toml
[project]
name = "gr-demo-3dgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=14.4.0",
    "gradio>=5.34.2",
    "groq>=0.28.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = ["ruff>=0.12.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
